services/alert-notification/requirements.txt
# Alert Notification Service Requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.19.0

# HTTP Client
httpx==0.26.0

# Email (SendGrid)
sendgrid==6.11.0

# SMS (Twilio)
twilio==8.11.0

# Async
asyncio==3.4.3  # Built-in

# Logging
structlog==24.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
